refactor(caching): consolidate caching onto aragora.caching substrate
Move the cache registry/protocol, adaptive TTL cache, and Redis-backed TTL
cache implementations into aragora/caching/ (registry.py, adaptive.py,
redis.py) and re-export the in-memory TTL primitives via aragora/caching/ttl.py.
aragora/cache.py, aragora/performance/adaptive_cache.py, and
aragora/utils/redis_cache.py become two-sided deprecation shims: plain import
works and re-exports the full public surface; importing under
-W error::DeprecationWarning fails; each warning names the old path. The three
distinct CacheStats APIs (decorator/registry/adaptive) are preserved.

Harden aragora/utils/redis_config.py lifecycle lock symmetry: close_redis_pool()
and reset_redis_state() now take _redis_lock (matching get_redis_pool's
first-use init guard) so teardown/reset cannot race a concurrent init on the
shared globals.

Internal non-handler caller aragora/performance/__init__.py and the moved-impl
tests migrate to the new homes; old paths keep working for external users.

Fulfills VAL-P4A-008.
