+++ tests/utils/test_redis_cache.py
@@ -21,7 +21,7 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from aragora.utils.redis_cache import (
+from aragora.caching.redis import (
     HybridTTLCache,
     RedisTTLCache,
 )
@@ -406,7 +406,7 @@ class TestRedisLazyInitialization:
     def test_get_redis_lazy_init(self):
         """_get_redis should lazily initialize on first call."""
         cache = RedisTTLCache(prefix="test")
-        with patch("aragora.utils.redis_cache.RedisTTLCache._get_redis") as mock_get_redis:
+        with patch("aragora.caching.redis.RedisTTLCache._get_redis") as mock_get_redis:
             mock_get_redis.return_value = None
             cache.get("key1")
             mock_get_redis.assert_called()
